try to type objs with only interior pointers
Previously, we were unable to produce a type for objects that only had
interior pointers, and no 0-offset pointers. This produced a lot of
"unk"-typed objects.

Now, if we have an object with just an interior pointer, we guess the
type of the main object by using the type of the interior pointer. I'm
not sure whether this has any major downsides.